chore: refactor competing proposals sidebar into smaller components
