+++ src/components/Proposal/View/Budget/AvailableOverBudgetBar.css
@@ -1,0 +1,43 @@
+.AvailableOverBudgetBar {
+  display: flex;
+  flex-direction: row;
+  align-items: stretch;
+  width: 100%;
+  border-radius: 4px;
+  position: absolute;
+}
+
+.AvailableOverBudget__Available {
+  background: var(--white-900);
+  border-top-left-radius: 2px;
+  border-bottom-left-radius: 2px;
+  height: 2px;
+  position: relative;
+  z-index: 10;
+  margin-top: 6px;
+  margin-left: 3px;
+}
+
+.AvailableOverBudgetBar__Allocated {
+  background: transparent !important;
+  height: 2px !important;
+}
+
+.AvailableOverBudgetBar__ArrowMarker {
+  position: relative;
+  overflow: visible !important;
+  right: 3px;
+  bottom: 4px;
+}
+
+.AvailableOverBudgetBar__HundredPercent {
+  font-style: normal;
+  font-weight: 400;
+  font-size: 13px;
+  line-height: 24px;
+  color: var(--black-600);
+  opacity: 0.45;
+  position: relative;
+  right: 28px;
+  top: 20px;
+}
