+++ src/components/Proposal/View/Budget/CompetingProposalsSidebar.css
@@ -27,161 +27,7 @@
   color: var(--black-800);
 }
 
-.ContestedBudgetCard__Title {
+.CompetingProposalsSidebar__CardRow {
   display: flex;
   justify-content: space-between;
-  width: 66%;
-}
-
-.ContestedBudgetCard {
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  gap: 8px;
-  margin-bottom: 2px;
-}
-
-.ContestedBudgetCard__Row {
-  display: flex;
-  justify-content: space-between;
-}
-
-.ContestedBudgetCard__Legend {
-  width: 8px;
-  height: 8px;
-  border-radius: 6px;
-  margin-right: 6px;
-}
-
-.ContestedLabel {
-  font-weight: 600;
-  font-size: 17px;
-  line-height: 18px;
-  letter-spacing: 0.4px;
-  color: var(--black-800);
-  margin-right: 6px;
-}
-
-.ContestedLegend {
-  background: rgb(var(--purple-raw-e));
-}
-
-.ContestedOverbudgetLegend {
-  background: var(--red-800);
-}
-
-.GrantedFundsLegend {
-  background: var(--black-600);
-}
-
-.UncontestedFundsLegend {
-  background: var(--alpha-black-300);
-}
-
-.AvailableOverBudgetLegend {
-  background: var(--white-900);
-  border: 2px solid var(--red-800);
-  width: 19px !important;
-  height: 6px !important;
-  border-radius: 5px;
-}
-
-.GrantedFundsLabel {
-  font-weight: 400;
-  font-size: 13px;
-  line-height: 18px;
-  color: var(--black-600);
-  text-transform: uppercase;
-  gap: 2px;
-  display: flex;
-  margin-right: 6px;
-}
-
-.GrantedFundsPercentageLabel {
-  font-weight: 400;
-  font-size: 13px;
-  color: var(--black-600);
-  line-height: 24px;
-  opacity: 0.45;
-}
-
-.GrantedFundsBar {
-  background: var(--black-600);
-  border-radius: 4px;
-  margin-right: 2px !important;
-}
-
-.GrantedFundsBarOverbudget {
-  height: 8px !important;
-}
-
-.CompetingProposalBudgetBar {
-  background: rgb(var(--purple-raw-e));
-  border-top-left-radius: 4px;
-  border-bottom-left-radius: 4px;
-}
-
-.CompetingProposalBudgetBar + .CompetingProposalBudgetBar {
-  border-top-left-radius: 0 !important;
-  border-bottom-left-radius: 0 !important;
-}
-
-.CompetingProposalBudgetBar--selected {
-  background: var(--gradient-purple-2);
-  box-shadow: 0 1px 4px rgb(var(--purple-raw-e));
-  z-index: 3;
-}
-
-.CompetingProposalOverbudgetBar {
-  background: var(--red-800);
-  height: 8px !important;
-  border-top-left-radius: 4px;
-  border-bottom-left-radius: 4px;
-}
-
-.CompetingProposalOverbudgetBar + .CompetingProposalOverbudgetBar {
-  border-top-left-radius: 0;
-  border-bottom-left-radius: 0;
-}
-
-.DistributionBarItem--selected.CompetingProposalOverbudgetBar--selected {
-  box-shadow: 0 1px 4px var(--purple-raw-b) !important;
-  border-radius: 10px !important;
-  height: 10px !important;
-}
-
-.ThisInitiativeBar {
-  background: var(--gradient-purple-1);
-  box-shadow: 0 1px 4px var(--alpha-fuchsia-500) !important;
-  height: 8px !important;
-  scale: 1.05 1;
-  border-radius: 10px !important;
-}
-
-.ThisInitiativeOverbudgetBar {
-  background: var(--red-800) !important;
-  border-radius: 10px;
-  scale: 1 !important;
-  box-shadow: 0 1px 4px var(--alpha-fuchsia-500) !important;
-}
-
-.DistributionBarItem.ThisInitiativeOverbudgetBar {
-  height: 14px !important;
-}
-
-.UncontestedBudgetBar {
-  background: var(--alpha-black-300);
-  border-radius: 4px;
-  margin-left: 2px;
-}
-
-.AvailableOverBudgetBar {
-  background: var(--white-900);
-  border-top-left-radius: 2px;
-  border-bottom-left-radius: 2px;
-  height: 2px;
-  position: relative;
-  z-index: 10;
-  margin-top: 6px;
-  margin-left: 3px;
 }
