+++ src/components/Proposal/View/Budget/ContestedBudgetDistributionBar.css
@@ -2,46 +2,3 @@
   display: flex;
   position: relative;
 }
-
-.ContestedBudgetBar {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  border-radius: 4px;
-  height: 6px;
-  min-width: 6px;
-  margin: 0;
-}
-
-.TransparentBar {
-  background: transparent !important;
-  height: 2px !important;
-}
-
-.AvailableOverBudgetDistributionBar {
-  display: flex;
-  flex-direction: row;
-  align-items: stretch;
-  width: 100%;
-  border-radius: 4px;
-  position: absolute;
-}
-
-svg.ArrowMarker {
-  position: relative;
-  overflow: visible;
-  right: 3px;
-  bottom: 4px;
-}
-
-.HundredPercent {
-  font-style: normal;
-  font-weight: 400;
-  font-size: 13px;
-  line-height: 24px;
-  color: var(--black-600);
-  opacity: 0.45;
-  position: relative;
-  right: 28px;
-  top: 20px;
-}
